Mover stepper ref to Mooc Slide
